Remove .vercelignore and move utils to utils dir
