--- .vercelignore
@@ -1,1 +1,0 @@
-/api/utils.go
