+++ api/index.go
@@ -1,68 +1,22 @@
 package api
 
 import (
-	"fmt"
 	"html/template"
 	"net/http"
-	"time"
+
+	"github.com/NDoolan360/go-htmx-site/utils"
 )
 
-type Index struct {
-	Title          string
-	Description    string
-	InternalLinks  []Link
-	ExternalLinks  []Link
-	ThemeSwitchSVG template.HTML
-	Profile
-	Experiences  []Experience
-	ToolSections []ToolSection
-	Copyright    string
-}
-
-type Link struct {
-	Label string
-	URL   template.URL
-	Logo  template.HTML
-}
-
-type Profile struct {
-	ImageAttr  []template.HTMLAttr
-	Paragraphs []string
-}
-
-type Experience struct {
-	Date
-	Workplace Link
-	Positions []Position
-	Topics    []string
-	Education bool
-}
-
-type Date struct {
-	Start string
-	End   string
-}
-
-type Position struct {
-	Title   string
-	Current bool
-}
-
-type ToolSection struct {
-	Title string
-	Links []Link
-}
-
 func GetIndex(w http.ResponseWriter, r *http.Request) {
-	tmp := template.Must(template.ParseFiles(
-		FromPWD("api/template/index.gohtml"),
-		FromPWD("api/template/head.gohtml"),
+	tmpl := template.Must(template.ParseFiles(
+		utils.FromPWD("api/template/index.gohtml"),
+		utils.FromPWD("api/template/head.gohtml"),
 	))
 
-	err := tmp.Execute(w, Index{
+	err := tmpl.Execute(w, utils.Index{
 		Title:       "Nathan Doolan",
 		Description: "A personal website showcasing Nathan Doolan's journey as a full-time software engineer in Melbourne. Explore his professional experience, projects, and interests in technology, board games, and 3D printing.",
-		InternalLinks: []Link{
+		InternalLinks: []utils.Link{
 			{
 				Label: "About",
 				URL:   "#about",
@@ -76,35 +30,35 @@ func GetIndex(w http.ResponseWriter, r *http.Request) {
 				URL:   "#projects",
 			},
 		},
-		ThemeSwitchSVG: IgnoreErr(GetSVGLogo("theme_switch")),
-		ExternalLinks: []Link{
+		ThemeSwitchSVG: utils.GetSVGLogo("theme_switch"),
+		ExternalLinks: []utils.Link{
 			{
-				"Github",
-				"https://github.com/NDoolan360",
-				IgnoreErr(GetSVGLogo("github")),
+				Label: "Github",
+				URL:   "https://github.com/NDoolan360",
+				Logo:  utils.GetSVGLogo("github"),
 			},
 			{
-				"LinkedIn",
-				"https://www.linkedin.com/in/nathan-doolan-835a13171",
-				IgnoreErr(GetSVGLogo("linkedin")),
+				Label: "LinkedIn",
+				URL:   "https://www.linkedin.com/in/nathan-doolan-835a13171",
+				Logo:  utils.GetSVGLogo("linkedin"),
 			},
 			{
-				"Discord",
-				"https://discord.com/users/nothindoin",
-				IgnoreErr(GetSVGLogo("discord")),
+				Label: "Discord",
+				URL:   "https://discord.com/users/nothindoin",
+				Logo:  utils.GetSVGLogo("discord"),
 			},
 			{
-				"Cults3D",
-				"https://cults3d.com/en/users/ND360",
-				IgnoreErr(GetSVGLogo("cults3d")),
+				Label: "Cults3D",
+				URL:   "https://cults3d.com/en/users/ND360",
+				Logo:  utils.GetSVGLogo("cults3d"),
 			},
 			{
-				"Boardgame Geek",
-				"https://boardgamegeek.com/user/Nothin_Doin",
-				IgnoreErr(GetSVGLogo("bgg")),
+				Label: "Boardgame Geek",
+				URL:   "https://boardgamegeek.com/user/Nothin_Doin",
+				Logo:  utils.GetSVGLogo("bgg"),
 			},
 		},
-		Profile: Profile{
+		Profile: utils.Profile{
 			ImageAttr: []template.HTMLAttr{
 				template.HTMLAttr(`title="Picture of me at the Strawberry Field in Liverpool"`),
 				template.HTMLAttr(`srcset="/images/profile-192.webp, /images/profile-512.webp 2.66x, /images/profile-792.webp 4.125x"`),
@@ -114,100 +68,91 @@ func GetIndex(w http.ResponseWriter, r *http.Request) {
 				"With a background in computer science, I'm always eager to learn new things, tackle new challenges, and find satisfaction in the simple joy of crafting an elegant solution to a problem.",
 			},
 		},
-		Experiences: []Experience{
+		Experiences: []utils.Experience{
 			{
-				Date: Date{"Jan 2024", "Present"},
-				Workplace: Link{
-					"Kaluza",
-					"https://kaluza.com/",
-					IgnoreErr(GetSVGLogo("kaluza")),
+				Date: utils.Date{Start: "Jan 2024", End: "Present"},
+				Workplace: utils.Link{
+					Label: "Kaluza",
+					URL:   "https://kaluza.com/",
+					Logo:  utils.GetSVGLogo("kaluza"),
 				},
-				Positions: []Position{{"Software Engineer", true}},
+				Positions: []utils.Position{{Title: "Software Engineer", Current: true}},
 				Topics:    []string{"Typescript", "Git", "Github Actions"},
 			},
 			{
-				Date: Date{"Jul 2021", "Dec 2023"},
-				Workplace: Link{
-					"Gentrack",
-					"https://gentrack.com/",
-					IgnoreErr(GetSVGLogo("gentrack")),
+				Date: utils.Date{Start: "Jul 2021", End: "Dec 2023"},
+				Workplace: utils.Link{
+					Label: "Gentrack",
+					URL:   "https://gentrack.com/",
+					Logo:  utils.GetSVGLogo("gentrack"),
 				},
-				Positions: []Position{
-					{"Intermediate Software Engineer", false},
-					{"Junior Software Engineer", false},
-					{"Graduate Software Engineer", false},
+				Positions: []utils.Position{
+					{Title: "Intermediate Software Engineer"},
+					{Title: "Junior Software Engineer"},
+					{Title: "Graduate Software Engineer"},
 				},
 				Topics: []string{"Git", "SQL", "Github Actions", "Docker", "Jenkins", "API Design", "Unit Testing"},
 			},
 			{
-				Date: Date{"Feb 2018", "Jul 2021"},
-				Workplace: Link{
-					"Proquip Rental & Sales",
-					"https://pqrs.com.au/",
-					IgnoreErr(GetSVGLogo("proquip")),
+				Date: utils.Date{Start: "Feb 2018", End: "Jul 2021"},
+				Workplace: utils.Link{
+					Label: "Proquip Rental & Sales",
+					URL:   "https://pqrs.com.au/",
+					Logo:  utils.GetSVGLogo("proquip"),
 				},
-				Positions: []Position{
-					{"IT Support Specialist", false},
-					{"IT/Marketing Assistant", false},
-					{"Administrative Assistant", false},
+				Positions: []utils.Position{
+					{Title: "IT Support Specialist"},
+					{Title: "IT/Marketing Assistant"},
+					{Title: "Administrative Assistant"},
 				},
 				Topics: []string{"IT Support", "Adobe Suite", "Social Media Marketing", "Wordpress", "Google Analytics"},
 			},
 			{
-				Date: Date{"Feb 2018", "Feb 2021"},
-				Workplace: Link{
-					"University of Melbourne",
-					"https://www.unimelb.edu.au/",
-					IgnoreErr(GetSVGLogo("melbourneuniversity")),
+				Date: utils.Date{Start: "Feb 2018", End: "Feb 2021"},
+				Workplace: utils.Link{
+					Label: "University of Melbourne",
+					URL:   "https://www.unimelb.edu.au/",
+					Logo:  utils.GetSVGLogo("melbourneuniversity"),
 				},
-				Positions: []Position{{"Bachelor of Science: Computing and Software Systems", false}},
+				Positions: []utils.Position{{Title: "Bachelor of Science: Computing and Software Systems"}},
 				Education: true,
 			},
 		},
-		ToolSections: []ToolSection{
+		ToolSections: []utils.ToolSection{
 			{
-				"Built with:",
-				[]Link{
+				Title: "Built with:",
+				Links: []utils.Link{
 					{
-						"htmx",
-						"https://htmx.org",
-						IgnoreErr(GetSVGLogo("htmx")),
+						Label: "htmx",
+						URL:   "https://htmx.org",
+						Logo:  utils.GetSVGLogo("htmx"),
 					},
 					{
-						"hyperscript",
-						"https://hyperscript.org",
-						IgnoreErr(GetSVGLogo("hyperscript")),
+						Label: "hyperscript",
+						URL:   "https://hyperscript.org",
+						Logo:  utils.GetSVGLogo("hyperscript"),
 					},
 					{
-						"Tailwind CSS",
-						"https://tailwindcss.com",
-						IgnoreErr(GetSVGLogo("tailwind")),
+						Label: "Tailwind CSS",
+						URL:   "https://tailwindcss.com",
+						Logo:  utils.GetSVGLogo("tailwind"),
 					},
 				},
 			},
 			{
-				"Deployed with:",
-				[]Link{
+				Title: "Deployed with:",
+				Links: []utils.Link{
 					{
-						"Vercel",
-						"https://vercel.com",
-						IgnoreErr(GetSVGLogo("vercel")),
+						Label: "Vercel",
+						URL:   "https://vercel.com",
+						Logo:  utils.GetSVGLogo("vercel"),
 					},
 				},
 			},
 		},
-		Copyright: Copyright("Nathan Doolan"),
+		Copyright: utils.Copyright("Nathan Doolan"),
 	})
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 	}
-}
-
-var Now = func() time.Time {
-	return time.Now()
-}
-
-func Copyright(name string) string {
-	year := Now().Year()
-	return fmt.Sprintf("© %s %d", name, year)
 }
